requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

matplotlib==3.10.8
pandas==3.0.2
python-dotenv==1.2.2
streamlit==1.56.0
openpyxl==3.1.5
langchain==1.2.15
langchain-openai==1.2.0
langchain-experimental==0.4.1